adding js for dedicated report refinement
