+++ static/js/form.js
@@ -45,61 +45,3 @@ platformSelector.addEventListener('change',(event)=>{
     storeForms(selection = event.target.value);
 });
 
-const ReportType = document.getElementById("reportType");
-const dateRangeHtml = document.getElementById("dateRange");
-const shipDateHtml = document.getElementById("shipDates")
-
-
-/* To make sure only a particular form will be affected */
-const reportForm = document.getElementById("reportForm");
-const formTitle = document.getElementById("formTitle");
-const submitButton = document.getElementById("submitButton");
-
-
-async function apiAccess(apiUrl){
-    try{
-        const response = await fetch(apiUrl);
-        if (!response.ok){
-            throw new Error(`Response : ${response.status}`);
-        }
-        const result = await response.json();
-        return result;
-    }
-    catch(error){
-        console.error(error.message);
-    }
-}
-const baseUrl = '/api/router/';
-
-
-let columnDiv = document.getElementById("columnCheckBoxes");
-
-ReportType.addEventListener("change",async ()=>{
-    columnDiv.innerHTML = ""; /*resetting report column selector*/
-    let profiles = await apiAccess(apiUrl = baseUrl + 'reports');
-    let columns;
-    profiles.forEach((profile) => {
-        if (profile["main_section"] === ReportType.value){
-            columns = profile["columns"].split(",");
-        }
-    });
-    
-    if (columns.length > 0){
-        columns.forEach((column) =>{
-            let columnLabel = document.createElement("label");
-            columnLabel.className = "form-check-label"; 
-            columnLabel.innerText = column;
-
-            let columnInput = document.createElement("input");
-            columnInput.className = "form-check-input";
-            columnInput.type = "checkbox";
-            columnInput.name = "report_column"; columnInput.value = column;
-
-            columnDiv.appendChild(columnLabel);
-            columnDiv.appendChild(columnInput);
-
-        });
-    }
-});
-
-
