+++ templates/connections.html
@@ -7,7 +7,7 @@
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.min.js" integrity="sha384-cVKIPhGWiC2Al4u+LWgxfKTRIcfu0JTxR+EQDz/bgldoEyl4H0zUF0QKbrJ0EcQF" crossorigin="anonymous"></script>
 
 <!-- Custom CSS -->
-<link rel="stylesheet" href="{% static 'css/common.css' %}">
+<link rel="stylesheet" href="{% static 'css/theme.css' %}">
 <link rel="stylesheet" href="{% static 'css/sidebar.css' %}">
 <link rel="stylesheet" href="{% static 'css/form.css' %}">
 <link rel="stylesheet" href="{% static 'css/dashboard.css' %}">
@@ -15,7 +15,9 @@
 <!-- Custom JS -->
 <script src="{% static 'js/debug.js' %}" defer></script>
 <script src="{% static 'js/form.js' %}" defer></script>
+<script src="{% static 'js/data_refinement.js' %}" defer></script>
 <script type="module" src="{% static 'js/api.js' %}" defer></script>
 
 <!-- HTMX -->
 <script src="https://cdn.jsdelivr.net/npm/htmx.org@2.0.6/dist/htmx.min.js"></script>
+data_refinement
